tests/requirements.txt
azure.iot.device
azure.eventhub
